+++ scripts/delayed.js
@@ -11,44 +11,11 @@ governing permissions and limitations under the License.
 */
 
 // eslint-disable-next-line import/no-cycle
-import { sampleRUM, loadScript } from './scripts.js';
-
-function loadAdobeLaunch() {
-  const { host } = window.location;
-  let scriptUrl = 'https://assets.adobedtm.com/52be42758645/f4f76a9081b6/launch-bd7785fa0dac-development.min.js';
-  if (/\.hlx\.page/.test(host) || /\.hlx\.live/.test(host)) {
-    // if on hlx.page or hlx.live upgrade to stage library
-    scriptUrl = 'https://assets.adobedtm.com/52be42758645/f4f76a9081b6/launch-37775755caa8-staging.min.js';
-  }
-
-  if (/rockstar.adobeevents.com/.test(host) || /main--aem-rockstar-website--adobe.hlx.live/.test(host)) {
-    // if on a prod domain, then use the prod library
-    scriptUrl = 'https://assets.adobedtm.com/52be42758645/f4f76a9081b6/launch-d4e94ba1de20.min.js';
-  }
-
-  loadScript(scriptUrl, null, null, true);
-}
-
-function instrumentPage() {
-  window.adobeDataLayer = window.adobeDataLayer || [];
-
-  let theme = '';
-  const themeEL = document.querySelector('meta[name="theme"]');
-  if (themeEL) {
-    theme = themeEL.getAttribute('content');
-  }
-
-  window.adobeDataLayer.push({
-    page: {
-      title: document.title,
-      theme,
-    },
-  });
-}
+import { sampleRUM } from './scripts.js';
+import { instrument } from './instrument.js';
 
 // Core Web Vitals RUM collection
 sampleRUM('cwv');
 
 // add more delayed functionality here
-instrumentPage();
-loadAdobeLaunch();
+instrument();
